Builder/Vector3 - double => float.

+++ Builder/Builder/Utilities/Tools/Vector/Vector3.h
@@ -5,5 +5,5 @@ struct Vector3
 	Vector3();
 	Vector3(double x, double y, double z);
 
-	double posX, posY, posZ;
+	float posX, posY, posZ;
 };
